Separated the take and put with an fsync
Separated the action of pulling the old Handle out of  the MVar
and putting a new one in.  I put an fsync in the middle
which runs when flagged.  This should force a lock on the
file handle and prevent things from disappearing before
the fsync gets a chance to do what it needs to.